Continuous threshold
Cleaned up the imports of state observers and transformers.
Gave the continuous transformers thresholds to allow for easy stopping.
Fixed the rank of a few actions in the state transformer tests.
Small cleanup to pydoc comments for state transformer tests.
Added an assert to a vector environment test to fix a type error.
Updated the continuous example environment to use the stop threshold.